requirements.txt
streamlit
tensorflow-cpu
haroun
opencv-python-headless
tensorflow_addons==0.13.0
torch
numpy



